Moving History Time Period logic into HistoryFetchers
